MPPA - The project compiles.
Supports very simple programs that load integer immediates. It starts
the main, loads integer in registers, and return correctly.

Addition in Mach not yet supported, but should not be hard to add them.

Function calls are not yet supported.

The ABI for now is the same as the RiscV, with a small twist: $ra is
first loaded in a user register, then this user register is pushed
(instead of pushing $ra straight away).